feat(challenges_list_screen): Add RefreshIndicator to challenges list screen
