+++ src/top_app/lib/core/router/app_route_names.dart
@@ -7,8 +7,8 @@ class AppRouteNames {
   static const String signUp = '/sign-up';
   static const String countdownRecord = '/countdown-record';
   static const String signIn = '/sign-in';
-  static const String showChallengesList = 'challenges';
-  static const String showChallengeDetail = '/challenge-detail';
+  static const String challengesList = 'challenges';
+  static const String challengeDetail = '/challenge-detail';
   static const String home = 'home';
   static const String navigation = '/';
   static const String submitActivityProof = '/submit-activity-proof';
