+++ src/top_app/lib/core/router/app_router.dart
@@ -4,8 +4,8 @@ import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 import 'package:top_app/modules/home/presentation/pages/home_screen.dart';
 import 'package:top_app/modules/navigation_bar/presentation/widgets/organisms/navigation_screen.dart';
-import 'package:top_app/modules/show_challenges_feature/sub_features/challenge_detail/presentation/widgets/screens/challenge_detail_screen.dart';
-import 'package:top_app/modules/show_challenges_feature/sub_features/show_challenges_list/presentation/widgets/screens/show_challenges_list_screen.dart';
+import 'package:top_app/modules/challenges/sub_features/challenge_detail/presentation/widgets/screens/challenge_detail_screen.dart';
+import 'package:top_app/modules/challenges/sub_features/challenges_list/presentation/widgets/screens/challenges_list_screen.dart';
 import 'package:top_app/modules/show_onboarding_feature/presentation/screens/ninety_nine_screen.dart';
 import 'package:top_app/modules/show_onboarding_feature/presentation/screens/start_or_level_up_screen.dart';
 import 'package:top_app/modules/show_onboarding_feature/presentation/screens/that_one_screen.dart';
@@ -138,7 +138,7 @@ class AppRouter extends RootStackRouter {
         ),
         AutoRoute(
           page: ChallengeDetailRoute.page,
-          path: AppRouteNames.showChallengeDetail,
+          path: AppRouteNames.challengeDetail,
         ),
         AutoRoute(
           initial: true,
@@ -150,8 +150,8 @@ class AppRouter extends RootStackRouter {
               path: AppRouteNames.home,
             ),
             AutoRoute(
-              page: ShowChallengesListRoute.page,
-              path: AppRouteNames.showChallengesList,
+              page: ChallengesListRoute.page,
+              path: AppRouteNames.challengesList,
             ),
           ],
         ),
