+++ src/top_app/lib/modules/challenges/sub_features/challenge_detail/domain/entities/competitor_info.dart
