+++ src/top_app/lib/modules/challenges/sub_features/challenge_detail/presentation/state_management/cubit/challenge_detail_cubit.freezed.dart
