+++ src/top_app/lib/modules/challenges/sub_features/challenge_detail/presentation/widgets/molecules/challenge_description.dart
