+++ src/top_app/lib/modules/challenges/sub_features/challenges_list/domain/repository/challenges_list_repository.dart
@@ -1,0 +1,6 @@
+import 'package:top_app/shared/entities/templates/challenge.dart';
+
+abstract class ChallengesListRepository {
+  /// Get all challenges from the database
+  Future<List<Challenge>> getChallenges();
+}
