--- src/top_app/lib/modules/show_challenges_feature/sub_features/challenge_detail/data/models/competitor_info_model.dart
@@ -1,19 +1,0 @@
-import 'package:top_app/modules/show_challenges_feature/sub_features/challenge_detail/domain/entities/competitor_info.dart';
-
-class CompetitorInfoModel {
-  final String userName;
-  final String profilePictureUrl;
-
-  CompetitorInfoModel({required this.userName, required this.profilePictureUrl});
-
-  factory CompetitorInfoModel.fromJson(Map<String, dynamic> json) {
-    return CompetitorInfoModel(
-      userName: json['userName'],
-      profilePictureUrl: json['userProfilePictureUrl'],
-    );
-  }
-
-  CompetitorInfo toEntity() {
-    return CompetitorInfo(userName: userName, profilePictureUrl: profilePictureUrl);
-  }
-}
