--- src/top_app/lib/modules/show_challenges_feature/sub_features/challenge_detail/domain/repository/challenge_detail_repository.dart
@@ -1,7 +1,0 @@
-import 'package:top_app/modules/show_challenges_feature/sub_features/challenge_detail/domain/entities/competitor_info.dart';
-
-abstract class ChallengeDetailRepository {
-  Future<List<CompetitorInfo>> getCompetitorsInfo(String challengeId);
-
-  Future<void> joinChallenge(String challengeId);
-}
