--- src/top_app/lib/modules/show_challenges_feature/sub_features/show_challenges_list/domain/repository/show_challenges_list_repository.dart
@@ -1,6 +1,0 @@
-import 'package:top_app/shared/entities/templates/challenge.dart';
-
-abstract class ShowChallengesListRepository {
-  /// Get all challenges from the database
-  Future<List<Challenge>> getChallenges();
-}
